tools: revert default tree for orphan files
The problem with below patch is that some files were with no git tree in
MAINTAINERS on purpose: some files are maintained in the main repository,
but we leave subtrees touch them like release notes or makefile/meson
files.

So reverting it, and adding a comment to document the reason.

Fixes: b338e13817dc ("tools: fix tree detection for orphan files")

Signed-off-by: David Marchand <david.marchand@redhat.com>
Acked-by: Thomas Monjalon <thomas@monjalon.net>